Added template to Vector2D to work with both floats and ints, updated code everywhere

+++ cppFallingSands-master/src/Elements/BaseElements/Moveable.cpp
@@ -16,7 +16,7 @@ void Moveable::updateVelocity(int& newCol, int upDown)
 	if (yVelocity > getTerminalY() * upDown) yVelocity = getTerminalY();
 }
 
-double Moveable::randomNumber()
+float Moveable::randomNumber()
 {
 	static std::default_random_engine rng(static_cast<unsigned>(std::chrono::system_clock::now().time_since_epoch().count()));
