+++ cppFallingSands-master/src/Elements/Smoke.cpp
@@ -4,8 +4,8 @@
 
 Smoke::Smoke()
 {
-    setIsGas(true);
-    setColor(SDL_MapRGBA(SDL_AllocFormat(SDL_PIXELFORMAT_RGBA32), 130 + randomNumber(), 130+randomNumber(), 130, 255));
+	setIsGas(true);
+	setColor(SDL_MapRGBA(SDL_AllocFormat(SDL_PIXELFORMAT_RGBA32), (Uint8)130.0f + randomNumber(), (Uint8)130.0f + randomNumber(), 130, 255));
 }
 
 Smoke::~Smoke() {}
